requirements/pip.txt
boto3==1.34.69
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-decouple==2.1
django-extra-fields
django-filter==23.2
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
num2words==0.5.12
numpy==2.0.0
openpyxl==3.1.2
pandas==2.2.2
psycopg2-binary==2.9.7
requests==2.28.1
xhtml2pdf==0.2.15
django-crontab==0.7.1
